Move CombinedResourceOwnerQuery to Security namespace
